requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
Faker==18.10.1
pytest==7.3.1
selenium==4.9.1
